Include some info about v2
Hi! Just saw you are working on v2 for Transformer. Can you please update this README or create a new one to explain how v2 is different, why is it useful, and whether it is a stable version to use (or when it will become such). Adding even few words about it would be very beneficial. Thank you!